requirements.txt
BzrTools==2.6.0
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
Pillow==2.3.0
PyJWT==1.6.4
SecretStorage==2.0.0
Werkzeug==0.14.1
argparse==1.2.1
bzr==2.7.0dev1
chardet==2.0.1
click==6.7
colorama==0.2.5
configobj==4.7.2
decorator==3.4.0
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
ipython==1.2.1
itsdangerous==0.24
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lpthw.web==1.1
matplotlib==1.3.1
mercurial==2.8.2
numpy==1.8.2
oauth==1.0.1
paramiko==1.10.1
pexpect==3.1
pycrypto==2.6.1
pygobject==3.12.0
pygpgme==0.3
pyparsing==2.0.1
python-apt==0.9.3.5ubuntu2
python-dateutil==1.5
pytz==2012c
requests==2.2.1
scipy==0.13.3
simplegeneric==0.8.1
simplejson==3.3.1
six==1.5.2
stevedore==0.14.1
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
